components moved to generator tests
